update habit saga to only listen for archived/unarchived

+++ src/redux-modules/habits/sagas/subcribeToHabitsSaga.js
@@ -13,13 +13,23 @@ import { firestore } from "../../../firebase/firebase";
 import habitsSlice from "../habitsSlice";
 import sessionSlice from "redux-modules/session/sessionSlice";
 import currentUserIdSelector from "redux-modules/session/selectors/currentUserIdSelector";
+import selectIsArchiveRoute from "../selectors/selectIsArchiveRoute";
+import { LOCATION_CHANGE } from "connected-react-router";
 
-function subscriptionChannel({ userId }) {
+function subscriptionChannel({ userId, archived }) {
   return eventChannel((emitter) => {
-    const habitsRef = firestore
-      .collection("habits")
-      .where("userId", "==", userId);
-    // .where('archived', '==', false);
+    let habitsRef;
+    if (archived) {
+      habitsRef = firestore
+        .collection("habits")
+        .where("userId", "==", userId)
+        .where("archived", "==", true);
+    } else {
+      habitsRef = firestore
+        .collection("habits")
+        .where("userId", "==", userId)
+        .where("archived", "==", false);
+    }
 
     const unsubscribe = habitsRef.onSnapshot((querySnapshot) => {
       querySnapshot.docChanges().forEach((change) => {
@@ -37,38 +47,49 @@ function subscriptionChannel({ userId }) {
 
 export function* listen(action) {
   const currentUserId = yield select(currentUserIdSelector);
+  const archived = yield select(selectIsArchiveRoute);
 
-  const channel = yield call(subscriptionChannel, { userId: currentUserId });
+  if (currentUserId) {
+    const channel = yield call(subscriptionChannel, {
+      userId: currentUserId,
+      archived,
+    });
 
-  try {
-    while (true) {
-      const { data: habit, id, type } = yield take(channel);
+    try {
+      while (true) {
+        const { data: habit, id, type } = yield take(channel);
 
-      switch (type) {
-        case "added": {
-          yield put(habitsSlice.actions.addHabit({ habit, id }));
-          break;
+        switch (type) {
+          case "added": {
+            yield put(habitsSlice.actions.addHabit({ habit, id }));
+            break;
+          }
+          case "modified": {
+            yield put(habitsSlice.actions.updateHabit({ habit, id }));
+            break;
+          }
+          case "removed": {
+            yield put(habitsSlice.actions.deleteHabit({ habit, id }));
+            break;
+          }
+          default:
+            break;
         }
-        case "modified": {
-          yield put(habitsSlice.actions.updateHabit({ habit, id }));
-          break;
-        }
-        case "removed": {
-          yield put(habitsSlice.actions.deleteHabit({ habit, id }));
-          break;
-        }
-        default:
-          break;
       }
+    } finally {
+      if (yield cancelled()) channel.close();
     }
-  } finally {
-    if (yield cancelled()) channel.close();
   }
 }
 
 export function* handleSubscribe(action) {
   const listenTask = yield fork(listen, action);
   const unsubscribed = yield take(sessionSlice.routines.logout.SUCCESS);
+  const routeChange = yield take(LOCATION_CHANGE);
+
+  if (routeChange.type === LOCATION_CHANGE) {
+    yield cancel(listenTask);
+  }
 
   if (unsubscribed.type === sessionSlice.routines.logout.SUCCESS) {
     yield cancel(listenTask);
@@ -81,6 +102,7 @@ function* subscribeToHabitsSaga() {
       sessionSlice.actions.setCurrentUser,
       sessionSlice.routines.login.SUCCESS,
       sessionSlice.routines.createUser.SUCCESS,
+      LOCATION_CHANGE,
     ],
     handleSubscribe,
   );
